fix(commerce): split server-only rate loader + work around zod v4 partial-refinement
Two build-time fixes for the 5-feature PR:

1. `lib/commerce/currency.ts` was importing `createAdminClient` (next/headers)
   yet being pulled into client bundles via <PriceDisplay>. Split the
   server-only `loadPygRates` into `currency-server.ts` (with `import
   'server-only'`) and kept the pure conversion helpers where the client
   component can safely reach them.

2. Zod v4 rejects `.partial()` on schemas with `.refine()`. Exported
   DiscountInputShape (unrefined object) so the PATCH route can derive a
   partial patch schema; full-body create still uses DiscountInputSchema
   with the refinement intact.

Verified: `npm run build` now passes.
